game/requirements.txt
pygame
pyvidplayer2
pillow
keyboard
pywin32
SpeechRecognition
SpeechRecognition[whisper-local]
easyocr
